+++ packages/android_intent/lib/android_intent.dart
@@ -14,7 +14,7 @@ const String kChannelName = 'plugins.flutter.io/android_intent';
 class AndroidIntent {
   /// Builds an Android intent with the following parameters
   /// [action] refers to the action parameter of the intent.
-  /// [flags] is the list of strings that will be converted to the flags. 
+  /// [flags] is the list of strings that will be converted to the flags.
   /// [category] refers to the category of the intent, can be null.
   /// [data] refers to the string format of the URI that will be passed to
   /// intent.
@@ -37,7 +37,7 @@ class AndroidIntent {
         _platform = platform ?? const LocalPlatform();
 
   final String action;
-  final List<String> flags;
+  final List<int> flags;
   final String category;
   final String data;
   final Map<String, dynamic> arguments;
@@ -45,6 +45,23 @@ class AndroidIntent {
   final String componentName;
   final MethodChannel _channel;
   final Platform _platform;
+
+  bool _isPowerOfTwo(int x) {
+    /* First x in the below expression is for the case when x is 0 */
+    return x != 0 && ((x & (x - 1)) == 0);
+  }
+
+  @visibleForTesting
+  int convertFlags(List<int> flags) {
+    int finalValue = 0;
+    for (int i = 0; i < flags.length; i++) {
+      if (!_isPowerOfTwo(flags[i])) {
+        throw ArgumentError.value(flags[i], 'flag\'s value must be power of 2');
+      }
+      finalValue |= flags[i];
+    }
+    return finalValue;
+  }
 
   /// Launch the intent.
   ///
@@ -54,7 +71,7 @@ class AndroidIntent {
     assert(_platform.isAndroid);
     final Map<String, dynamic> args = <String, dynamic>{'action': action};
     if (flags != null) {
-      args['flags'] = flags;
+      args['flags'] = convertFlags(flags);
     }
     if (category != null) {
       args['category'] = category;