+++ packages/google_maps_flutter/lib/src/google_map.dart
@@ -192,7 +192,7 @@ class _GoogleMapState extends State<GoogleMap> {
   @override
   Widget build(BuildContext context) {
     final Map<String, dynamic> creationParams = <String, dynamic>{
-      'initialCameraPosition': widget.initialCameraPosition?._toMap(),
+      'initialCameraPosition': widget.initialCameraPosition?.toMap(),
       'options': _googleMapOptions.toMap(),
       'markersToAdd': _serializeMarkerSet(widget.markers),
       'polygonsToAdd': _serializePolygonSet(widget.polygons),
@@ -298,6 +298,14 @@ class _GoogleMapState extends State<GoogleMap> {
     final MarkerId markerId = MarkerId(markerIdParam);
     if (_markers[markerId]?.onTap != null) {
       _markers[markerId].onTap();
+    }
+  }
+
+  void onMarkerDragEnd(String markerIdParam, LatLng position) {
+    assert(markerIdParam != null);
+    final MarkerId markerId = MarkerId(markerIdParam);
+    if (_markers[markerId]?.onDragEnd != null) {
+      _markers[markerId].onDragEnd(position);
     }
   }